requirements.txt
numexpr
langchain-core
langchain-openai
chromadb
google-generativeai
python-dotenv